requirements.txt
pysrt==1.1.2
requests==2.32.3
pymongo==4.13.0
python-dotenv==1.1.0
qdrant-client
